requirements.txt
wikipedia
stop-words
Pillow
requests
wordcloud
stop-words
